pyproject.toml
[tool.poetry]
name = "mdrepaintclothing"
version = "0.1.0"
description = "服装图像处理一体化平台基础工程"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.13"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
flake8 = "^6.0.0"
isort = "^5.0.0"
pre-commit = "^3.0.0"
coverage = "^7.0.0"

[tool.poetry.group.dev.dependencies]
black = "22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 